Remove redundant signal blocking in signal handler
Because Linux manual signal(7) says 'Any signals specified in
act->sa_mask when registering the handler with sigprocmask(2) are
added to the thread's signal mask. The signal being delivered is also
added to the signal mask, unless SA_NODEFER was specified when
registering the handler. These signals are thus blocked while the
handler executes.'